Reorganised the project structure
- Separated the executor and scheduler into their own modules
- Moved the static files to a separate module
- Updated the import paths accordingly
