--- scraper/classes/__init__.py
@@ -1,4 +1,0 @@
-"""
-Classes package for the scraper. This package contains
-the main classes that are used across the scraper module.
-"""
