+++ scraper/static/paths.py
