+++ scraper/static/seed_urls.py
